cputools finished

+++ src/c/cputools/memusage.h
@@ -1,4 +1,4 @@
 unsigned long atolu(char *buf);
 unsigned long GetMemTotal();
 unsigned long GetProcessMemTotal(int pid);
-int GetProcessMemUsage(int pid);
+float GetProcessMemUsage(int pid);
